CI/reflector/requirements.txt
# Dependencies for the reflection code generator
libclang>=16.0.0
Jinja2>=3.1.0
